move argument parser to own file
